removed the folder structure and removed some codes from common function.js
