+++ common/functions.js
@@ -21,6 +21,8 @@ import {
   ipaRequestApi,
   ipaSuccessHandler,
   documentUpload,
+  checkMode,
+  customSetFocus,
 } from '../creditcards/corporate-creditcardFunctions.js';
 
 import {
@@ -31,7 +33,7 @@ import {
 import fetchAuthCode from './idcomutil.js';
 
 import {
-  urlPath, santizedFormDataWithContext, getTimeStamp, formUtil, clearString,
+  urlPath, santizedFormDataWithContext, getTimeStamp, clearString,
 } from './formutils.js';
 
 import {
@@ -47,75 +49,6 @@ const { currentFormContext } = corpCreditCardContext;
 
 // dynamically we can change according to journey
 const journeyNameConstant = JOURNEY_NAME;
-/**
- * @name checkMode - check the location
- * @param {object} globals -
- * @return {PROMISE}
- */
-function checkMode(globals) {
-  debugger;
-  const formData = globals.functions.exportData();
-  const idcomVisit = formData?.queryParams?.authmode; // "DebitCard"
-  const aadharVisit = formData?.queryParams?.visitType; // "EKYC_AUTH
-  // temporarly added referenceNumber check for IDCOMM redirection to land on submit screen.
-  if (aadharVisit === 'EKYC_AUTH' && formData?.aadhaar_otp_val_data?.message && formData?.aadhaar_otp_val_data?.message === 'Aadhaar OTP Validate success') {
-    globals.functions.setProperty(globals.form.corporateCardWizardView, { visible: true });
-    globals.functions.setProperty(globals.form.otpPanel, { visible: false });
-    globals.functions.setProperty(globals.form.loginPanel, { visible: false });
-    globals.functions.setProperty(globals.form.getOTPbutton, { visible: false });
-    globals.functions.setProperty(globals.form.consentFragment, { visible: false });
-    globals.functions.setProperty(globals.form.welcomeText, { visible: false });
-    const {
-      result: {
-        Address1, Address2, Address3, City, State, Zipcode,
-      },
-    } = formData.aadhaar_otp_val_data;
-    const {
-      executeInterfaceReqObj: {
-        requestString: {
-          officeAddress1, officeAddress2, officeAddress3, officeCity, officeState, officeZipCode,
-          communicationAddress1, communicationAddress2, communicationAddress3, communicationCity, communicationState, comCityZip,
-        },
-      },
-    } = formData.currentFormContext;
-    const aadharAddress = [Address1, Address2, Address3, City, State, Zipcode]?.filter(Boolean)?.join(', ');
-    const officeAddress = [officeAddress1, officeAddress2, officeAddress3, officeCity, officeState, officeZipCode]?.filter(Boolean)?.join(', ');
-    const communicationAddress = [communicationAddress1, communicationAddress2, communicationAddress3, communicationCity, communicationState, comCityZip]?.filter(Boolean)?.join(', ');
-    const { AddressDeclarationAadhar, addressDeclarationOffice, CurrentAddressDeclaration } = globals.form.corporateCardWizardView.confirmAndSubmitPanel.addressDeclarationPanel;
-    globals.functions.setProperty(AddressDeclarationAadhar.aadharAddressSelectKYC, { value: aadharAddress });
-    globals.functions.setProperty(addressDeclarationOffice.officeAddressSelectKYC, { value: officeAddress });
-    globals.functions.setProperty(CurrentAddressDeclaration.currentResidenceAddress, { value: communicationAddress });
-  } if (idcomVisit === 'DebitCard') {
-    const resultPanel = formUtil(globals, globals.form.resultPanel);
-    resultPanel.visible(false);
-    globals.functions.setProperty(globals.form.otpPanel, { visible: false });
-    globals.functions.setProperty(globals.form.loginPanel, { visible: false });
-    globals.functions.setProperty(globals.form.getOTPbutton, { visible: false });
-    globals.functions.setProperty(globals.form.consentFragment, { visible: false });
-    globals.functions.setProperty(globals.form.welcomeText, { visible: false });
-    globals.functions.setProperty(globals.form.resultPanel.successResultPanel, { visible: false });
-    globals.functions.setProperty(globals.form.resultPanel.errorResultPanel, { visible: false });
-    globals.functions.setProperty(globals.form.confirmResult, { visible: false });
-    const userRedirected = true;
-    executeInterfacePostRedirect('idCom', userRedirected, globals);
-  }
-}
-
-/**
- * does the custom show hide of panel or screens in resend otp.
- * @param {string} errorMessage
- * @param {number} numRetries
- * @param {object} globals
- */
-function customSetFocus(errorMessage, numRetries, globals) {
-  if (typeof numRetries === 'number' && numRetries < 1) {
-    globals.functions.setProperty(globals.form.otpPanel, { visible: false });
-    globals.functions.setProperty(globals.form.submitOTP, { visible: false });
-    globals.functions.setProperty(globals.form.resultPanel, { visible: true });
-    globals.functions.setProperty(globals.form.resultPanel.errorResultPanel, { visible: true });
-    globals.functions.setProperty(globals.form.resultPanel.errorResultPanel.errorMessageText, { value: errorMessage });
-  }
-}
 
 /**
  * generates the otp
